NEW:V1.2.3 新增支持并验证转矩模式，但是老问题BISSC 9s左右死掉100%复现
连接上位机验证了切换到转矩模式后可以启动运行，但是BISSC获取的老问题100%必现
Reset
SPI不管用，且逻辑分析仪抓取BISS-C波形失真严重，看起来不是SPI外设的问题，需要复位AMG2000

+++ Core/Inc/global_data.h
@@ -174,13 +174,13 @@ typedef struct {
 
 typedef struct {
   uint8_t g_posi[4];      // unit depends on BISS-C
-  // 反馈值
+  // 上报反馈值
 	int32_t g_Distance; 	// um
 	int32_t g_Speed; 			// rpm
-  int16_t g_Torque;     // N.m
+  int16_t g_Torque;     // N.m 
   int16_t g_phaseAmp;   // A
+  //log显示
   float g_realTimeTorque; // 实时转矩
-
 	uint32_t g_InitialPosi; //um
 
   volatile uint8_t g_DS402_SMStatus;  //状态字处于0x1237时，此状态为1
@@ -266,7 +266,7 @@ typedef struct {
     uint32_t EHeader;        // Ethernet帧头
     uint32_t ENum;           // Ethernet帧号
     uint32_t ELen;           // Ethernet长度 Byte
-    uint8_t EType;           // 报文类型
+    uint16_t EType;           // 报文类型
     uint8_t CASNodeID;       // 数据来源
     uint8_t curWorkMode;     // 当前工作模式
     uint16_t statusWord;     // 状态字
